+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * Support for multiple web clients: `curl` (default) and `wget`.
 * Migrated environment variable configuration to `loda/setup.txt`.
+* Improved messages and usability of the setup command.